Moved env config into seperate file
